refactor: useConstant
