+++ site/src/composables/constant/variables/colors.ts
@@ -1,0 +1,8 @@
+export const PRESET_COLORS = [
+  "#000000",
+  "#377bb5",
+  "#ca3832",
+  "#ee8732",
+  "#9c5bde",
+  "#43912b"
+];
