+++ site/src/composables/constant/variables/papers.ts
@@ -1,0 +1,12 @@
+export const PAPER_SIZES = {
+  A4: {
+    h: 297 + 2,
+    w: 210
+  },
+  letter: {
+    h: 279.4 + 3,
+    w: 215.9
+  }
+};
+
+export const MM_TO_PX = 3.78;
