+++ site/src/composables/constant/variables/render.ts
@@ -1,0 +1,3 @@
+export const PRINT_BOTTOM = 10;
+export const PREVIEW_ID = "preview";
+export const PREVIEW_SELECTOR = `#resume-${PREVIEW_ID}`;
