+++ site/src/composables/index.ts
@@ -1,3 +1,4 @@
 export { useDataStore } from "./stores/data";
 export { useStyleStore } from "./stores/style";
 export { useMonaco } from "./monaco";
+export { useConstant, type Font, type ValidPaperSize } from "./constant";
