+++ site/src/utils/css.ts
@@ -1,4 +1,5 @@
 import { injectCss } from "@ohmycv/dynamic-css";
+import { useConstant } from "~/composables/constant";
 import type { ResumeStyles } from "~/composables/stores/style";
 
 const { RENDER } = useConstant();
